+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Added analyzer `COIMBRA0014`.
 - Added analyzer `COIMBRA0015`.
 - Added analyzer `COIMBRA0016`.
+- Added analyzer `COIMBRA0017`.
+- Added analyzer `COIMBRA0018`.
 
 ## [10.3.1] - 2022-05-25